requirements.txt
asgiref==3.8.1
bcc==0.1.10
Django==5.1.7
django-filter==24.3
django-geojson==4.2.0
django-leaflet==0.31.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
numpy==2.2.3
psycopg2-binary==2.9.10
sqlparse==0.5.3
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.12.2
